+++ torchao/sparsity/__init__.py
@@ -7,12 +7,10 @@
 from .wanda import WandaSparsifier  # noqa: F403
 from .utils import PerChannelNormObserver  # noqa: F403
 from .sparse_api import apply_sparse_semi_structured, apply_fake_sparsity
-from .dynamic_quant_sparse import Int8DynamicallyQuantizedSemiStructuredSparseLinearWeight
 
 __all__ = [
     "WandaSparsifier",
     "PerChannelNormObserver",
     "apply_sparse_semi_structured",
     "apply_fake_sparsity",
-    "Int8DynamicallyQuantizedSemiStructuredSparseLinearWeight",
 ]
